requirements.txt
h5py
PyYAML
pandas
torch
torchvision
pillow
torchsummary